moved InstanceValidationService to a service package
